+++ survey-training-core/src/main/java/net/atos/survey/core/dao/QuestionDao.java
@@ -9,8 +9,8 @@ import net.atos.survey.core.entity.Question;
 
 
 @Local
-public interface QuestionDao extends Dao<Integer, Question> {
+public interface QuestionDao extends Dao<Long, Question> {
 
-
+	
 	
 }
