+++ survey-training-core/src/main/java/net/atos/survey/core/dao/SurveyDao.java
@@ -9,7 +9,7 @@ import net.atos.survey.core.entity.Survey;
 
 
 @Local
-public interface SurveyDao extends Dao<Integer, Survey> {
+public interface SurveyDao extends Dao<Long, Survey> {
