+++ survey-training-core/src/main/java/net/atos/survey/core/dao/SurveyTemplateDao.java
@@ -9,7 +9,7 @@ import net.atos.survey.core.entity.SurveyTemplate;
 
 
 @Local
-public interface SurveyTemplateDao extends Dao<Integer, SurveyTemplate> {
+public interface SurveyTemplateDao extends Dao<Long, SurveyTemplate> {
