+++ survey-training-core/src/main/java/net/atos/survey/core/dao/TrainingDao.java
@@ -9,7 +9,7 @@ import net.atos.survey.core.entity.Training;
 
 
 @Local
-public interface TrainingDao extends Dao<Integer, Training> {
+public interface TrainingDao extends Dao<Long, Training> {
